split out implementation files from cs_Logger.h
