+++ source/include/logging/impl/cs_LogNone.h
@@ -1,0 +1,8 @@
+/**
+ * Only to be directly included in cs_Logger.h.
+ *
+ * This implementation just swallows its arguments.
+ */
+
+#define _log(level, addNewLine, fmt, ...)
+#define _logArray(level, addNewLine, pointer, size, ...)
